requirements.txt
asgiref==3.4.1
certifi==2020.12.5
Django==3.2.8
django-haystack==3.1.1
django-mdeditor==0.1.18
django-redis==5.0.0
django-taggit==2.0.0
djangorestframework==3.12.4
importlib-metadata==4.8.2
jieba==0.42.1
Markdown==3.3.6
Pillow==8.4.0
Pygments==2.10.0
PyMySQL==1.0.2
pytz==2021.3
redis==3.5.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
Whoosh==2.7.4
zipp==3.6.0
asgiref==3.4.1
certifi==2020.12.5
Django==3.2.8
django-haystack==3.1.1
django-mdeditor==0.1.18
django-redis==5.0.0
django-taggit==2.0.0
djangorestframework==3.12.4
importlib-metadata==4.8.2
jieba==0.42.1
Markdown==3.3.6
Pillow==8.4.0
Pygments==2.10.0
PyMySQL==1.0.2
pytz==2021.3
redis==3.5.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
Whoosh==2.7.4
zipp==3.6.0
